% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer_prep.R
\name{officer_prep}
\alias{officer_prep}
\title{function to re-factor anonymous officer values by values present in given dataset}
\usage{
officer_prep(df)
}
\arguments{
\item{df}{df with officer_anon and officer_order columns}
}
\value{
df with officer names ordered by order
}
\description{
function to re-factor anonymous officer values by values present in given dataset
}
