% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_prep.R
\name{race_prep}
\alias{race_prep}
\title{Order race values before plot}
\usage{
race_prep(
  df,
  race_col = "race",
  race_vector_order = c("Black", "Hispanic", "White", "Asian", "Other",
    "Native American")
)
}
\arguments{
\item{df}{a stops or arrests dataframe with a race column}

\item{race_col}{a column for race/ethnicity values in the arrest/stops dataframe}

\item{race_vector_order}{character vector of how to order race-values appearing in dataframe in visualizations}
}
\value{
df with race column ordered as factor
}
\description{
Function to order race consistently for graph colors. Dataframe must have a race column with values matching those in race_vector_order. Should run before running plotly.
}
