% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizer.R
\name{summarizer}
\alias{summarizer}
\title{Function to generate group-counts and percentages}
\usage{
summarizer(df, group_cols, output_prefixs)
}
\arguments{
\item{df}{Data-frame with each row corresponding to an observation you want counted by group.}

\item{group_cols}{Data-masked list of columns to group the dataframe by, in order of order you want the grouping done.}

\item{output_prefixs}{Character-vector of equal length as the group_cols with prefixes for output columns corresponding to each group_col; e.g., "type" if you're grouping by type will produce type_num and type_pct.}
}
\value{
A summary dataframe with information on the number of observations per each group/combination of groups in order. E.g., c(animal, breed, likeability) will group by animal, then animal/breed, then animal/breed/likeability, and calculate the percentage of each animal group of all animals, of each animal/breed group of each animal group, and of each animal/breed/likeability group of each animal/breed group.
}
\description{
Function to generate group-counts and percentages
}
