% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_race_recode.R
\name{ucr_race_recode}
\alias{ucr_race_recode}
\title{ucr_race_recode}
\usage{
ucr_race_recode(string)
}
\arguments{
\item{string}{a string for the UCR race value}
}
\value{
A string with recoded race values
}
\description{
Recode UCR race values to more usable values
}
