% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ward_prep.R
\name{ward_prep}
\alias{ward_prep}
\title{Order race values before plot}
\usage{
ward_prep(
  df,
  ward_col = "a.ward",
  ward_vector_order = c("Ward 1", "Ward 2", "Ward 3", "Ward 4", "Ward 5", "Ward 6",
    "OOC", "UNK")
)
}
\arguments{
\item{df}{a stops or arrests dataframe with a ward column}

\item{ward_col}{a column for ward values in the arrest/stops dataframe}

\item{ward_vector_order}{character vector of how to order wards in visualizations/charts}
}
\value{
df with race column ordered as factor
}
\description{
Function to order race consistently for graph colors. Dataframe must have a race column with values matching those in ward_vector_order. Should run before running plotly.
}
