% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_lookup.R
\name{age_lookup}
\alias{age_lookup}
\title{Recode age to range}
\usage{
age_lookup(num, traffic = T)
}
\arguments{
\item{num}{a numeric age value}

\item{traffic}{Whether working with traffic stops data or other datasets}
}
\value{
string
}
\description{
Recodes numeric age value to age range
}
\examples{
age_lookup(16)
}
