% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_lookup.R
\name{race_lookup}
\alias{race_lookup}
\title{Function to convert ACS races to police-tracked races}
\usage{
race_lookup(string, traffic = T)
}
\arguments{
\item{string}{representing acs races}
}
\value{
recoded string
}
\description{
Function to convert ACS races to police-tracked races
}
